requirements.txt
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
asyncpg==0.30.0
click==8.1.7
databases==0.9.0
exceptiongroup==1.2.2
fastapi==0.115.6
ruff>=0.2.0
mypy==1.9.0
h11==0.14.0
idna==3.10
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.32.1
requests == 2.31.0
docker==7.1.0
elasticsearch>=9.0.0
urllib3==2.2.3
python-dotenv>=1.0.0
aiohttp==3.11.10
redis==5.0.0
celery>=5.3.0
flower==2.0.1
greenlet==3.1.1
# LLM Service Dependencies
pytesseract==0.3.10
Pillow==11.2.1
geopandas==0.14.3
fiona==1.9.5
openai==1.12.0
pytest>=8.2.0
pytest-cov==4.1.0
pytest-asyncio==0.25.3